requirements.txt
langchain
streamlit
openai
tiktoken
python-dotenv
langchain-community